requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
distro==1.9.0
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.9.0
lxml==5.3.1
numpy==2.2.4
ollama==0.4.7
openai==1.68.0
pillow==11.1.0
pip==25.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyPDF2==3.0.1
python-docx==1.1.2
requests==2.32.3
setuptools==75.8.0
sniffio==1.3.1
sounddevice==0.5.1
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
wheel==0.45.1
